requirements.txt
pyyaml
dotty-dict